\name{G.asTable}
\alias{G.asTable}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
G.asTable
}
\description{
Returns kcl.waterloo.data.GJDataTable given an R matrix as input.
}
\usage{
G.asTable(inMatrix)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{inMatrix}{
An R matrix
}
}


\author{
Malcolm Lidierth
}


\seealso{
G.asList
}




% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{programming}
\keyword{Interfaces to Other Languages}% __ONLY ONE__ keyword per line
