\name{GShell.clear}
\alias{GShell.clear}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
GShell.clear
}
\description{
Clears the Groovy shell's context delting all variable.
}
\usage{
GShell.clear()
}


\value{
Does not return a value.
}

\author{
Malcolm Lidierth
}

\examples{
example <- function(){
print("This demo does nothing unless you uncomment the GShell.clear() call")
# Uncomment the line below to get this script to work
#GShell.clear()
}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{programming}
\keyword{Interfaces to Other Languages}

