\name{GShell.console}
\alias{GShell.console}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
GShell.console
}
\description{
Creates and displays a Groovy Console sharing the shell of the GShell sigleton instance.
}
\usage{
GShell.console()
}
%- maybe also 'usage' for other objects documented here.

\value{
Does not return a value.
}

\author{
%%  ~~who you are~~
Malcolm Lidierth
}


\examples{
example <-function(input){
GShell.console()
}

## The function is currently defined as
function () 
J("kcl/waterloo/shell/GShell", "console")
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{programming}
\keyword{Interfaces to Other Languages}% __ONLY ONE__ keyword per line
