\name{GShell.editor}
\alias{GShell.editor}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
GShell.editor
}
\description{
Creates and displays a basic editor for creating Groovy/Java scripts.
}
\usage{
GShell.editor()
}

\author{
%%  ~~who you are~~
Malcolm Lidierth
}

\examples{
example <-function(){
GShell.editor()
}
}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{programming}
\keyword{Interfaces to Other Languages}% __ONLY ONE__ keyword per line
