\name{GShell.evalEDTWith}
\alias{GShell.evalEDTWith}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
GShell.evalEDTWith
}
\description{
Evaluates the supplied script in the shell using the supplied shell context on the EDT.
}
\usage{
GShell.evalEDTWith(string, map)
}

\arguments{
\item{string}{
The code to evaluate.
}
\item{map}{
The map to use.
}
}

\author{
%%  ~~who you are~~
Malcolm Lidierth
}

\examples{
example <-function(string){
GShell.evalEDTWith(string)
}
}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{programming}
\keyword{Interfaces to Other Languages}% __ONLY ONE__ keyword per line
