\name{GShell.getScriptList}
\alias{GShell.getScriptList}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
GShell.getScriptList
}
\description{
Returns the  script list.
}
\usage{
GShell.getScriptList()
}



\author{
%%  ~~who you are~~
Malcolm Lidierth
}

\examples{
example <-function(){
GShell.getScriptList()
}
}



% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{programming}
\keyword{Interfaces to Other Languages}% __ONLY ONE__ keyword per line
