\name{GShell.getSource}
\alias{GShell.getSource}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
GShell.getSource
}
\description{
Returns the current script as a GroovyCodeSource instance.
}
\usage{
GShell.getSource()
}


\author{
%%  ~~who you are~~
Malcolm Lidierth
}

\examples{
example <-function(){
GShell.getSource()
}

## The function is currently defined as
function (string) 
J("kcl/waterloo/shell/GShell", "getSource")
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{programming}
\keyword{Interfaces to Other Languages}% __ONLY ONE__ keyword per line
