\name{GShell.getSwing}
\alias{GShell.getSwing}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
GShell.getSwing
}
\description{
Returns the SwingBuilder instance associated with GShell.
}
\usage{
GShell.getSwing()
}



\author{
%%  ~~who you are~~
Malcolm Lidierth
}


\examples{
example <-function(){
GShell.getSwing()
}
}



% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{programming}
\keyword{Interfaces to Other Languages}% __ONLY ONE__ keyword per line
