\name{GShell.load}
\alias{GShell.load}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
GShell.load
}
\description{
Loads a script and makes it the current script.
}
\usage{
GShell.load(script)
}

\arguments{
\item{script}{
The code source.
}
}


\author{
%%  ~~who you are~~
Malcolm Lidierth
}


\examples{
example <-function(script){
GShell.load(script)
}
}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{programming}
\keyword{Interfaces to Other Languages}% __ONLY ONE__ keyword per line
