\name{GShell.run}
\alias{GShell.run}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
GShell.run
}
\description{
Evaluates the current script in the shell using the current shell context and on the current thread.
}
\usage{
GShell.run()
}


\author{
%%  ~~who you are~~
Malcolm Lidierth
}

\examples{
example <-function(){
GShell.run()
}

## The function is currently defined as
function () 
J("kcl/waterloo/shell/GShell", "run")
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{programming}
\keyword{Interfaces to Other Languages}% __ONLY ONE__ keyword per line
