\name{GShell.runEDTWith}
\alias{GShell.runEDTWith}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
GShell.runEDTWith
}
\description{
Runs the current script in the shell using the supplied shell context on the EDT.
}
\usage{
GShell.runEDTWith(map)
}


\arguments{
\item{map}{
The map to use.
}
}

\author{
%%  ~~who you are~~
Malcolm Lidierth
}


\examples{
example <-function(map){
GShell.runEDTWith(map)
}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{programming}
\keyword{Interfaces to Other Languages}% __ONLY ONE__ keyword per line
