\name{GShell.setSource}
\alias{GShell.setSource}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
GShell.setSource
}
\description{
Returns the current script as a GroovyCodeSource instance.
}
\usage{
GShell.setSource(script)
}

\arguments{
\item{script}{
The GrrovyCodeSource instance to use as the current script.
}
}


\author{
%%  ~~who you are~~
Malcolm Lidierth
}


\examples{
example <-function(script){
GShell.setSource(script)
}
}



% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{programming}
\keyword{Interfaces to Other Languages}% __ONLY ONE__ keyword per line
